Fix AO_SIZE_MAX definition (Linux/musl-gcc) (Cherry-pick commit 7a99c0f from 'master' branch.)
* src/atomic_ops_malloc.c [SIZE_MAX && !CPPCHECK] (AO_SIZE_MAX): Cast
SIZE_MAX to size_t explicitly (to workaround incorrect definition
of SIZE_MAX - e.g. it is defined as ~0ULL in musl-tools which is not
correct for a 32-bit target).

+++ src/atomic_ops_malloc.c
@@ -140,7 +140,8 @@ static char *get_mmaped(size_t sz)
 # include <limits.h>
 #endif
 #ifdef SIZE_MAX
-# define AO_SIZE_MAX SIZE_MAX
+# define AO_SIZE_MAX ((size_t)SIZE_MAX)
+            /* Extra cast to workaround some buggy SIZE_MAX definitions. */
 #else
 # define AO_SIZE_MAX (~(size_t)0)
 #endif
